parsed input can be FullForm or InputForm...
I know this seems silly, since right now there is only a FullForm
output and FullForm -> FullForm should be a no-op (but sometimes it isn't).

However in the long term, I hope there are other kinds of output like
sage.

The FoxySheep{Parser,Listener,Visitor} modules and classes have been
renamed to InputForm since that is Mathematica's terminology and that
pairs with the name FullForm better.